Add stub functions and tests for new ReprEntry lines formatting.
Currently, a ValidationError is reported using a DatatestReprEntry
instance--a subclass of ReprEntry. But some report-handling functions
in pytest require a ReprEntry instance and will fail when given a
subclass (see shawnbrown/datatest#51). This will be addressed by
refactoring the plugin to modify ReprEntry.lines from the
pytest_runtest_logreport hook.

Refactoring this behavior will start with the following functions:

* _find_validationerror_start()
* _format_reprentry_lines()
* pytest_runtest_logreport()